Change the stream header format and reduce overhead

+++ internal/multiplex/frame.go
@@ -5,6 +5,6 @@ import ()
 type Frame struct {
 	StreamID uint32
 	Seq      uint32
-	Closing  uint32
+	Closing  uint8
 	Payload  []byte
 }
